ref(tiller): refactor tests in logical files
TestRunReleaseTest -> tiller/release_testing_test.go
TestGetReleaseContent -> tiller/release_content_test.go
TestGetReleaseStatusDeleted -> tiller/release_status_test.go
TestGetReleaseStatus -> tiller/release_status_test.go
TestRollbackReleaseFailure -> tiller/release_rollback_test.go
TestRollbackReleaseNoHooks -> tiller/release_rollback_test.go
TestRollbackRelease -> tiller/release_rollback_test.go
TestUninstallReleaseWithKeepPolicy -> tiller/release_uninstall_test.go
TestUninstallPurgeDeleteRelease -> tiller/release_uninstall_test.go
TestUninstallReleaseNoHooks -> tiller/release_uninstall_test.go
TestUninstallPurgeRelease -> tiller/release_uninstall_test.go
TestUninstallRelease -> tiller/release_uninstall_test.go
